Move login pages and remove unused imports
