+++ packages/frontend/src/components/user-page/register-page.js
