+++ packages/frontend/src/services/axios/token-service.js
@@ -19,7 +19,6 @@ const getLocalRefreshToken = () => {
   };
   
   const setUser = (user) => {
-    console.log(JSON.stringify(user));
     localStorage.setItem("user", JSON.stringify(user));
   };
