+++ packages/frontend/src/views/user-page/user-page.jsx
@@ -1,13 +1,11 @@
 import { useState } from 'react';
-import { PageContainer } from '../../components/utils/containers';
-import { FormGroup, FormFloatingInput, FormInput } from '../../components/utils/form';
 import { ViewPage } from '../../components/utils/pages'
 import { useDispatch, useSelector } from 'react-redux';
 import { userLogin, userRegister } from '../../redux/slices/user-state-slice';
-import { navigateTo, usePathString } from '../../hooks/location-path-hooks';
-import RegisterPage from './register-page';
-import LoginPage from './login-page';
-import RegisterChildPage from './register-child-page';
+import { usePathString } from '../../hooks/location-path-hooks';
+import RegisterPage from '../../components/user-page/register-page';
+import LoginPage from '../../components/user-page/login-page';
+import RegisterChildPage from '../../components/user-page/register-child-page';
 
 const UserPage = () => {
     const [getUser, setUser] = useState({
